fix: duration and price type

+++ createTables.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS movies_info(
 	id BIGSERIAL PRIMARY KEY,
 	name VARCHAR(50) UNIQUE NOT NULL,
 	description TEXT,
-	duration BIGINT NOT NULL,
-	price BIGINT NOT NULL
+	duration INTEGER NOT NULL,
+	price INTEGER NOT NULL
 );
